+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Infinite tasks are now shown as infinite in the task screen
 * Added support for more than the default matter types shown in the terminal
 * Fixed Identification Chamber not consuming the item when the chip storage place on top filled a chip
+* Fixed pipes not invalidating their power capabilities when the network changes
 
 # 1.2.1